+++ NEWS.md
@@ -10,6 +10,7 @@ This may have been addressed at 64a308ba50b9119108e8946737460de5997b805b by addi
 
 # vcfR 1.5.0.9000
 Released on CRAN 201X-XX-XX.
+* `extract.indels()` ignores GATK's <NON_REF>.
 * Added support for chromR objects with no gt slot to `proc.chromR()`.
 * Created `peak_to_ploid()` to call peaks and calculate dfe from `freq_peak()` output.
 * Created `freq_peak_plot()` to help visualize the output of `freq_peak()`.